refactor: MessageControllerAPI
