+++ src/main/java/org/dnlkk/controller/api/MessageControllerAPI.java
@@ -1,4 +1,4 @@
-package org.dnlkk.controller;
+package org.dnlkk.controller.api;
 
 import com.dnlkk.controller.annotations.*;
 import com.dnlkk.controller.responses.ResponseEntity;
